refactor(ServicesModels): #408 lift Services value types + namespace anchor out of Services
New SPM target `ServicesModels` mirrors the SearchModels /
SampleIndexModels / CorePackageIndexingModels split pattern. Lifts
the Services namespace anchor and four pure value types out of the
behavioural Services target so MCP and CLI surfaces can hold them
without depending on the full Services + Search + SampleIndex graph.

## New target: ServicesModels

Deps: [SearchModels, SharedCore, SharedConstants, SharedModels].
Foundation-layer only. Hosts:

- `Services` namespace anchor (with Services.Formatter sub-namespace
  and its HIG / Frameworks / Unified / Footer sub-sub-namespaces)
- `Services.SearchQuery` (struct, 11 fields, common search params)
- `Services.SearchFilters` (struct, 8 fields + hasActiveFilters)
- `Services.HIGQuery` (struct, 4 fields, HIG-specific query)
- `Services.Formatter.Config` (struct, 6 fields + cliDefault /
  mcpDefault / shared statics)

## File moves

Sources/Services/Services.swift -> Sources/ServicesModels/Services.swift
Sources/Services/Formatters/Services.Formatter.Config.swift
  -> Sources/ServicesModels/Services.Formatter.Config.swift
Plus new files Services.SearchQuery.swift, Services.SearchFilters.swift,
Services.HIGQuery.swift in ServicesModels.

The original homes (Services.SearchService.swift,
Services.HIGSearchService.swift) drop the lifted struct definitions
and keep their protocol + actor declarations; they gain
`import ServicesModels`.

## Consumer updates

All Services-target files extending the Services namespace add
`import ServicesModels` (the namespace anchor moved). CLI files
that use Services.SearchQuery / SearchFilters / HIGQuery /
Formatter.Config / ServiceContainer also gain the import.
SearchToolProvider gains the import. SearchToolProviderTests +
ServicesTests test targets gain the import (and the test target
deps in Package.swift gain ServicesModels).

## Package.swift

- New servicesModelsTarget + servicesModelsTestsTarget
- Services target gains ServicesModels dep
- SearchToolProvider target gains ServicesModels dep
- CLI target gains ServicesModels dep
- ServicesTests + SearchToolProviderTests gain ServicesModels dep

## Verification

xcrun swift build: clean
xcrun swift test: 1441 tests in 161 suites pass

## #408 progress

SearchToolProvider still imports Services for the behavioural
surface (DocsSearchService, TeaserService, UnifiedSearchService
actors + the formatter family). Dropping that requires a service-
as-protocol lift across all four service actors — separate multi-PR
slice. This PR is the foundation for that work: value types and
config types are now in a foundation-only target that consumers can
import without dragging in the full Services behavior.
