+++ Packages/Sources/CLI/Commands/CLI.Command.Search.SmartReport.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import ArgumentParser
 import Foundation
 import Logging
