+++ Packages/Sources/CLI/SearchModuleAlias.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 import Search
 import SearchModels
