+++ Packages/Sources/SearchToolProvider/CompositeToolProvider.swift
@@ -4,6 +4,7 @@ import MCPSharedTools
 import SampleIndexModels
 import SearchModels
 import Services
+import ServicesModels
 import SharedConstants
 import SharedCore
 import SharedUtils
