+++ Packages/Sources/Services/Formatters/Sample.Format.Markdown.Search.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 import SampleIndex
 import SampleIndexModels
