+++ Packages/Sources/Services/Formatters/Services.Formatter.Footer.Kind.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 
 // MARK: - Footer Kind
