+++ Packages/Sources/Services/Formatters/Services.Formatter.Footer.Search.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 import SharedConstants
