+++ Packages/Sources/Services/Formatters/Services.Formatter.Frameworks.Text.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 import SharedConstants
 import SharedCore
