+++ Packages/Sources/Services/Formatters/Services.Formatter.HIG.JSON.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 import SearchModels
 import SharedCore
