+++ Packages/Sources/Services/Formatters/Services.Formatter.HIG.Markdown.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 import SearchModels
 import SharedConstants
