+++ Packages/Sources/Services/Formatters/Services.Formatter.JSON.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 import SearchModels
 import SharedCore
