+++ Packages/Sources/Services/Formatters/Services.Formatter.Markdown.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 import SearchModels
 import SharedConstants
