+++ Packages/Sources/Services/Formatters/Services.Formatter.StringExtensions.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 
 // MARK: - Formatter String Utilities
