+++ Packages/Sources/Services/Formatters/Services.Formatter.TeaserResults.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 import SampleIndex
 import SampleIndexModels
