+++ Packages/Sources/Services/Formatters/Services.Formatter.Text.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 import SearchModels
 import SharedConstants
