+++ Packages/Sources/Services/ReadCommands/Services.UnifiedSearchService.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 import SampleIndex
 import SampleIndexModels
