+++ Packages/Sources/Services/Services.ServiceContainer.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import Foundation
 import SampleIndex
 import SearchModels
