+++ Packages/Tests/SearchToolProviderTests/CupertinoSearchToolProviderTests.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import TestSupport
 import SharedConstants
 import Testing
