+++ Packages/Tests/ServicesTests/ServicesTests.swift
@@ -1,3 +1,4 @@
+import ServicesModels
 import SharedConstants
 import Testing
 import Foundation
